arm64: configs: karate|karatep: add CONFIG_PCI_QUIRKS back
CONFIG_PCI_QUIRKS is mostly for PC use, and per https://lkml.org/lkml/2008/8/19/83
"This patch adds the CONFIG_PCI_QUIRKS option which allows to remove
all the PCI quirks, not necessarly used on embedded systems when PCI
is working properly. As this is a size-reduction option, it depends on
CONFIG_EMBEDDED. It allows to save almost 12 kilobytes of kernel code"
However removing it causing QCOM WiFi PCIe power regression.

Bug: 35710702
Test: marlin boot
Change-Id: I776cdd8a5d81ca4b13c42789c8e1bf59bbfd5ab9
Signed-off-by: Wei Wang <wvw@google.com>